backend/requirements.txt
# Chef's Kiss Backend - Python Age 5.0
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database & Supabase
supabase==2.3.4
sqlalchemy==2.0.25

# Caching
redis==5.0.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Image Processing
pillow==10.2.0

# Utilities
python-dotenv==1.0.0
